add vearch gpu, try to fix vearch filter =
